pyproject.toml
[project]
name = "embeddingbuddy"
version = "0.2.0"
description = "A Python Dash application for interactive exploration and visualization of embedding vectors through dimensionality reduction techniques."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dash>=2.17.1",
    "plotly>=5.17.0",
    "pandas>=2.1.4",
    "numpy>=1.24.4",
    "scikit-learn>=1.3.2",
    "dash-bootstrap-components>=1.5.0",
    "umap-learn>=0.5.8",
    "numba>=0.56.4",
    "openTSNE>=1.0.0",
    "mypy>=1.17.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.4.1",
    "pytest-cov>=4.1.0",
]
lint = [
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
security = [
    "bandit[toml]>=1.7.5",
    "safety>=2.3.0",
    "pip-audit>=2.6.0",
]
dev = [
    "embeddingbuddy[test,lint,security]",
]
all = [
    "embeddingbuddy[test,lint,security]",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
